[#161336913]modifies file structure
